requirements.txt
procgen==0.10.4
gym3==0.3.3
gym==0.18.3
torch==1.9.0
pytorch-lightning==1.2.5
torchvision==0.10.0
einops==0.3.0
cma==3.1.0
hydra-core==1.1.0
wandb==0.12.1
matplotlib==3.4.2
numpy==1.20.3
Pillow==8.2.0

